+++ src/main/java/ru/hse/roguelike/model/characters/CharacterType.java
@@ -1,4 +1,4 @@
-package ru.hse.roguelike.model.Characters;
+package ru.hse.roguelike.model.characters;
 
 /**
  * Enum class for character types.
