+++ src/main/java/ru/hse/roguelike/model/characters/Obstacle.java
@@ -1,4 +1,4 @@
-package ru.hse.roguelike.model.Characters;
+package ru.hse.roguelike.model.characters;
 
 import java.util.Random;
