+++ src/main/java/ru/hse/roguelike/model/characters/Points.java
@@ -1,4 +1,4 @@
-package ru.hse.roguelike.model.Characters;
+package ru.hse.roguelike.model.characters;
 
 /**
  * Represents points.
