+++ src/main/java/ru/hse/roguelike/model/levelbuilder/FromFileLevelBuilder.java
@@ -2,20 +2,20 @@ package ru.hse.roguelike.model.levelbuilder;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
-import ru.hse.roguelike.model.Characters.CharacterType;
-import ru.hse.roguelike.model.Characters.Empty;
-import ru.hse.roguelike.model.Characters.mob.Mob;
-import ru.hse.roguelike.model.Characters.GameCharacter;
-import ru.hse.roguelike.model.Characters.Inventory;
-import ru.hse.roguelike.model.Characters.Obstacle;
-import ru.hse.roguelike.model.Characters.Player;
-import ru.hse.roguelike.model.Characters.Points;
-import ru.hse.roguelike.model.Characters.Shelter;
+import ru.hse.roguelike.model.characters.CharacterType;
+import ru.hse.roguelike.model.characters.Empty;
+import ru.hse.roguelike.model.characters.mob.Mob;
+import ru.hse.roguelike.model.characters.GameCharacter;
+import ru.hse.roguelike.model.characters.Inventory;
+import ru.hse.roguelike.model.characters.Obstacle;
+import ru.hse.roguelike.model.characters.Player;
+import ru.hse.roguelike.model.characters.Points;
+import ru.hse.roguelike.model.characters.Shelter;
 import ru.hse.roguelike.model.Coordinates;
-import ru.hse.roguelike.model.Characters.mob.factory.MobFactory;
+import ru.hse.roguelike.model.characters.mob.factory.MobFactory;
 import ru.hse.roguelike.model.InventoryItem;
 import ru.hse.roguelike.model.Level;
-import ru.hse.roguelike.model.Characters.mob.factory.RedMobFactory;
+import ru.hse.roguelike.model.characters.mob.factory.RedMobFactory;
 
 import java.io.File;
 import java.io.IOException;
