+++ src/main/java/ru/hse/roguelike/model/levelbuilder/RandomLevelBuilder.java
@@ -1,25 +1,25 @@
 package ru.hse.roguelike.model.levelbuilder;
 
-import ru.hse.roguelike.model.Characters.CharacterType;
-import ru.hse.roguelike.model.Characters.Empty;
-import ru.hse.roguelike.model.Characters.mob.Mob;
-import ru.hse.roguelike.model.Characters.GameCharacter;
-import ru.hse.roguelike.model.Characters.Inventory;
-import ru.hse.roguelike.model.Characters.Obstacle;
-import ru.hse.roguelike.model.Characters.Player;
-import ru.hse.roguelike.model.Characters.Points;
-import ru.hse.roguelike.model.Characters.Shelter;
+import ru.hse.roguelike.model.characters.CharacterType;
+import ru.hse.roguelike.model.characters.Empty;
+import ru.hse.roguelike.model.characters.mob.Mob;
+import ru.hse.roguelike.model.characters.GameCharacter;
+import ru.hse.roguelike.model.characters.Inventory;
+import ru.hse.roguelike.model.characters.Obstacle;
+import ru.hse.roguelike.model.characters.Player;
+import ru.hse.roguelike.model.characters.Points;
+import ru.hse.roguelike.model.characters.Shelter;
 import ru.hse.roguelike.model.Coordinates;
-import ru.hse.roguelike.model.Characters.mob.factory.MobFactory;
+import ru.hse.roguelike.model.characters.mob.factory.MobFactory;
 import ru.hse.roguelike.model.InventoryItem;
 import ru.hse.roguelike.model.Level;
-import ru.hse.roguelike.model.LevelCharacteristics.FifthLevelCharacteristic;
-import ru.hse.roguelike.model.LevelCharacteristics.FirstLevelCharacteristic;
-import ru.hse.roguelike.model.LevelCharacteristics.FourthLevelCharacteristic;
-import ru.hse.roguelike.model.LevelCharacteristics.LevelCharacteristic;
-import ru.hse.roguelike.model.LevelCharacteristics.SecondLevelCharacteristic;
-import ru.hse.roguelike.model.LevelCharacteristics.ThirdLevelCharacteristic;
-import ru.hse.roguelike.model.Characters.mob.factory.YellowMobFactory;
+import ru.hse.roguelike.model.levelCharacteristics.FifthLevelCharacteristic;
+import ru.hse.roguelike.model.levelCharacteristics.FirstLevelCharacteristic;
+import ru.hse.roguelike.model.levelCharacteristics.FourthLevelCharacteristic;
+import ru.hse.roguelike.model.levelCharacteristics.LevelCharacteristic;
+import ru.hse.roguelike.model.levelCharacteristics.SecondLevelCharacteristic;
+import ru.hse.roguelike.model.levelCharacteristics.ThirdLevelCharacteristic;
+import ru.hse.roguelike.model.characters.mob.factory.YellowMobFactory;
 
 import java.util.HashMap;
 import java.util.List;
