+++ src/test/java/ru/hse/roguelike/model/ConfusedMobTest.java
@@ -2,10 +2,10 @@ package ru.hse.roguelike.model;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import ru.hse.roguelike.model.Characters.CharacterType;
-import ru.hse.roguelike.model.Characters.mob.Mob;
-import ru.hse.roguelike.model.Characters.mob.decorator.ConfusedMobDecorator;
-import ru.hse.roguelike.model.Characters.mob.strategy.AggressiveMobStrategy;
+import ru.hse.roguelike.model.characters.CharacterType;
+import ru.hse.roguelike.model.characters.mob.Mob;
+import ru.hse.roguelike.model.characters.mob.decorator.ConfusedMobDecorator;
+import ru.hse.roguelike.model.characters.mob.strategy.AggressiveMobStrategy;
 
 public class ConfusedMobTest {
     private Coordinates addCoordinates(Coordinates coordinates, Coordinates shift) {
