+++ src/test/java/ru/hse/roguelike/model/LevelBuilderTests.java
@@ -2,8 +2,8 @@ package ru.hse.roguelike.model;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import ru.hse.roguelike.model.Characters.Player;
-import ru.hse.roguelike.model.LevelCharacteristics.*;
+import ru.hse.roguelike.model.characters.Player;
+import ru.hse.roguelike.model.levelCharacteristics.*;
 import ru.hse.roguelike.model.levelbuilder.FromFileLevelBuilder;
 import ru.hse.roguelike.model.levelbuilder.LevelBuilder;
 import ru.hse.roguelike.model.levelbuilder.RandomLevelBuilder;
